v1.0.3.5
Changes to keep in sync with updates to Worldweep's living world rules. Most of the listed changes are a reflection of those alterations.

Note about changed feats: Feats on your character sheet are unique "items" that will not be changed retroactively. To get the new versions, delete the current versions off your character sheet and reapply the same feat in the same slot (or edit your existing one to match the new one, if you prefer).

Shardbound Feats:
* Slightly cleaner flavor text for Brilliant Rain.
* Clarified in Focused Recollection that it has to be a lore skill relevant to the creature targetted. NOTE: the Focused Recollection feat is a new item, and this is the feat which the effect links back to. If you don't replace it, the version on your character sheet and the version which the effect references won't agree. Save yourself the confusion and update the feat on your character sheet.
* Shardbound Weapon now takes 10 minutes to shape but remains indefinitely afterwards.
* Shardflesh now works with the unarmed attacks from the Shardbound Vampire and Werecreature archetypes.
* Shardlings cannot select familiar or master abilities that require them to use movement types they do not have.

Shardbound Vampire:
* Shardbound Vampire Dedication, Basic Shard Metamorphosis, Evolved Shard Metamorphosis: Replaced all references to "positive" and "negative" healing or damage with vitality and void healing or damage.
* Added a link to the Shardbound Vampire Archetype journal page to Shardbound Vampire Dedication, which was previously missing it.